Removed irrelevant files
