+++ main.py
@@ -1,174 +1,140 @@
 from dataclasses import dataclass, field
-from typing import List
+from datetime import datetime
+from typing import Dict, List
 
-from fastapi import FastAPI, Request, WebSocket
+from fastapi import FastAPI, WebSocket, WebSocketDisconnect
 from fastapi.templating import Jinja2Templates
+
+from utils import utils
 
 app = FastAPI()
 
 templates = Jinja2Templates(directory="static")
-# app.mount("/static", StaticFiles(directory="static"), name="static")
-
-html = """
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Chat</title>
-    </head>
-    <body>
-        <h1>WebSocket Chat</h1>
-        <h2>Your ID: <span id="ws-id"></span></h2>
-        <form action="" onsubmit="sendMessage(event)">
-            <input type="text" id="messageText" autocomplete="off"/>
-            <button>Send</button>
-        </form>
-        <ul id='messages'>
-        </ul>
-        <script>
-            var client_id = Date.now()
-            document.querySelector("#ws-id").textContent = client_id;
-            var ws = new WebSocket(`ws://localhost:8000/ws/${client_id}`);
-            ws.onmessage = function(event) {
-                var messages = document.getElementById('messages')
-                console.log(messages)
-                var message = document.createElement('li')
-                var content = document.createTextNode(event.data)
-                message.appendChild(content)
-                messages.appendChild(message)
-            };
-            function sendMessage(event) {
-                var input = document.getElementById("messageText")
-                ws.send(input.value)
-                input.value = ''
-                event.preventDefault()
-            }
-        </script>
-    </body>
-</html>
-"""
 
 
-class ConnectionManager:
-    """Connects, disconnects, can send individual messages to a socket, or broaddcasts messages."""
+@dataclass
+class User:
+    """User class to send messages to its `connection`."""
 
-    def __init__(self):
-        self.active_connections: List[WebSocket] = []
-        self.rooms: List[Room] = []
+    name: str
+    id: int
+    current_room: "Room"
+    connection: WebSocket
+
+    async def send_message(self, message: object):
+        """Sends `message` to its connection."""
+        await self.connection.send_json(
+            message
+        )
+
+
+@dataclass
+class Room:
+    """Room that hosts users as websocket connections."""
+
+    room_id: str
+    users: List[User] = field(default_factory=list)
+    owner: User = None
+    messages: List[Dict] = field(default_factory=dict)
 
     async def connect(self, websocket: WebSocket):
-        """Adds `websocket` to all active connections."""
+        """Connects the user defined `websocket` and sends a join to all other users in the room."""
         await websocket.accept()
-        self.active_connections.append(websocket)
 
-    def disconnect(self, websocket: WebSocket):
-        """Removes `websocket` from all active connections."""
-        self.active_connections.remove(websocket)
+        user_data = await websocket.receive_json()
+        await websocket.send_json(
+            {
+                "type": "room_join_success",
+                "room_id": self.room_id,
+                "username": user_data["name"],
+                "users": [i.name for i in self.users]
+            }
+        )
 
-    async def send_personal_message(self, message: str, websocket: WebSocket):
-        """Sends `message` to `websocket`."""
-        await websocket.send_text(message)
+        await self.send_all(
+            {
+                "type": "user_join",
+                "username": user_data["name"],
+                "time": datetime.utcnow().timestamp(),
+                "users": [i.name for i in self.users]
+            }
+        )
 
-    async def broadcast(self, message: str):
-        """Broadcasts message to all active connections."""
-        for connection in self.active_connections:
-            await connection.send_text(message)
+        self.users.append(User(name=user_data["name"], id=user_data["id"], current_room=self, connection=websocket))
+
+    async def disconnect(self, websocket: WebSocket):
+        """Removes the user defined in `websocket` from the room."""
+        user_to_remove: User = list(filter(lambda x: x.connection.client == websocket.client, self.users))[0]
+        self.users.remove(user_to_remove)
+
+        await self.send_all(
+            {
+                "type": "room_disconnect_success",
+                "username": user_to_remove.name,
+                "room_id": self.room_id,
+                "time": datetime.utcnow().timestamp(),
+                "users": [i.name for i in self.users]
+            }
+        )
+
+    async def send_all(self, message: object):
+        """Sends `message` to all users in the room."""
+        for user in self.users:
+            await user.send_message(message)
+
+
+@dataclass
+class ConnectionManager:
+    """Manager that holds a list of all the open rooms."""
+
+    rooms: List[Room] = field(default_factory=list)
+
+    async def broadcast(self, message: object):
+        """Sends `message` to all the rooms."""
+        for room in self.rooms:
+            await room.send_all(message)
 
     async def locate_room(self, room_id: str):
-        """Returns the room with id `room_id`."""
+        """Returns the room with id `room_id` from self.rooms."""
         for i in self.rooms:
             if room_id == i.room_id:
                 return i
 
         return None
 
+    async def locate_user(self, client_id: int):
+        """Returns the user with id `client_id`."""
+        for i in self.rooms:
+            for user in i.users:
+                if client_id == user.id:
+                    return user
 
-@dataclass
-class Room:
-    """Room to connect, disconnect, and broadcast messages to users."""
-
-    name: str
-    room_id: str
-    people: List[WebSocket] = field(default_factory=list)
-    owner: WebSocket = None
-
-    async def connect(self, websocket: WebSocket):
-        """Adds `websocket` to list of people."""
-        await websocket.accept()
-        self.people.append(websocket)
-
-    def disconnect(self, websocket: WebSocket):
-        """Removes `websocket` from list of people."""
-        self.people.remove(websocket)
-
-    async def broadcast(self, message: str):
-        """Sends a message to all users."""
-        for connection in self.people:
-            await connection.send_text(message)
+        return None
 
 
 manager = ConnectionManager()
 
 
-@app.get("/")
-async def get(request: Request):
-    """Todo (Firestar)"""
-    data = {"num": "not connected"}
-    return templates.TemplateResponse("index.html", {"request": request, "data": data})
+@app.websocket("/ws/{room_id}/{client_id}")
+async def chat_room(websocket: WebSocket, room_id: str, client_id: str):
+    """TODO: (firestar): write docstring"""
+    room = await manager.locate_room(room_id=room_id)
 
+    if room is None:
+        room = Room(room_id=room_id, users=[], messages=[])
+        manager.rooms.append(room)
 
-@app.websocket("/ws/{room_id}/{client_id}/{joining}")
-async def make_room(websocket: WebSocket, room_id: str, client_id: int, joining: bool):
-    """Creates a room."""
-    find = await manager.locate_room(room_id=room_id)
-    if not find:
-        if joining:
-            await manager.connect(websocket)
-            await manager.send_personal_message("not found", websocket)
-            return
-
-        else:
-            for i in manager.rooms:
-                if websocket in i.people:
-                    print("removing", i.name)
-                    if len(i.people) > 1:
-                        i.disconnect(websocket)
-
-                    else:
-                        manager.rooms.remove(i)
-
-            room = Room(name=str(client_id) + "'s room", room_id=room_id)
-            room.owner = websocket
-            manager.rooms.append(room)
-            await room.connect(websocket)
-            await room.broadcast(f"{client_id} has joined #{room_id}")
-            find = room
-
-    elif websocket not in find.people:
-        for i in manager.rooms:
-            if websocket in i.people:
-                print("removing", i.name)
-                if len(i.people) > 1:
-                    i.disconnect(websocket)
-
-                else:
-                    manager.rooms.remove(i)
-        if joining:
-            await manager.connect(websocket)
-            await manager.send_personal_message("found", websocket)
-            manager.disconnect(websocket)
-
-        else:
-            await find.connect(websocket)
-            await find.broadcast(f"{client_id} has joined #{find.room_id}")
+    await room.connect(websocket)
 
     try:
         while True:
-            data = await websocket.receive_text()
-            await find.broadcast(f"{client_id}: {data}")
+            data = await websocket.receive_json()
+            if data["type"] == "message_sent":
+                data["content"] = utils.scramble_sentence(data["content"])
+                room.messages.append({"user": data["username"], "content": data["content"]})
+                await room.send_all(data)
 
-    except Exception:
-        find.disconnect(websocket)
-        await find.broadcast(f"{client_id} has left the room")
-
-        if len(find.people) == 0:
-            manager.rooms.remove(find)
+    except WebSocketDisconnect:
+        await room.disconnect(websocket)
+        if len(room.users) == 0:
+            manager.rooms.remove(room)
